# api/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Polyad API",
    "description": "API pour le système d'IA autonome Polyad",
    "version": "1.0.0",
    "contact": {
      "email": "contact@polyad.ai"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Serveur principal"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "error"
          },
          "message": {
            "type": "string",
            "example": "Description de l'erreur"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["username", "password"],
        "properties": {
          "username": {
            "type": "string",
            "example": "admin"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        }
      },
      "SystemStatus": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          },
          "system": {
            "type": "object",
            "properties": {
              "cpu": {
                "type": "object",
                "properties": {
                  "usage": {
                    "type": "number",
                    "example": 45.2
                  },
                  "cores": {
                    "type": "integer",
                    "example": 8
                  }
                }
              },
              "memory": {
                "type": "object",
                "properties": {
                  "total": {
                    "type": "integer",
                    "example": 16384
                  },
                  "used": {
                    "type": "integer",
                    "example": 8192
                  },
                  "percent": {
                    "type": "number",
                    "example": 50
                  }
                }
              },
              "disk": {
                "type": "object",
                "properties": {
                  "total": {
                    "type": "integer",
                    "example": 512000
                  },
                  "used": {
                    "type": "integer",
                    "example": 256000
                  },
                  "percent": {
                    "type": "number",
                    "example": 50
                  }
                }
              },
              "temperature": {
                "type": "number",
                "example": 42
              }
            }
          },
          "agent": {
            "type": "object",
            "properties": {
              "initialized": {
                "type": "boolean",
                "example": true
              },
              "capabilities": {
                "type": "object",
                "additionalProperties": {
                  "type": "boolean"
                },
                "example": {
                  "Vision": true,
                  "Audio": true,
                  "Apprentissage": true,
                  "Actions": true
                }
              },
              "performance": {
                "type": "object",
                "additionalProperties": {
                  "type": "number"
                },
                "example": {
                  "accuracy": 85,
                  "speed": 78,
                  "efficiency": 82,
                  "learning": 90,
                  "adaptation": 75
                }
              }
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "summary": "Authentification et génération de token JWT",
        "tags": ["Authentification"],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentification réussie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "Requête invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentification échouée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Vérification de l'état de l'API",
        "tags": ["Système"],
        "security": [],
        "responses": {
          "200": {
            "description": "État de l'API",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "online"
                    },
                    "timestamp": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "version": {
                      "type": "string",
                      "example": "1.0.0"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/system/status": {
      "get": {
        "summary": "Obtenir l'état du système",
        "tags": ["Système"],
        "responses": {
          "200": {
            "description": "État du système",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemStatus"
                }
              }
            }
          },
          "401": {
            "description": "Non authentifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erreur serveur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}